# src/apiserver/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PulseOps V2 API",
    "version": "2.0.0",
    "description": "Enterprise platform API for PulseOps V2. Supports database management, authentication (JWT + HttpOnly cookies), module hot-dropping, and system configuration. All URLs are defined in urls.json — this spec references relative paths under the /api prefix.",
    "contact": {
      "name": "PulseOps Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative to host — works on any environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained via POST /auth/login. Expiry: 24h."
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "HttpOnly cookie set automatically on login. Same JWT as BearerAuth."
      }
    }
  },
  "paths": {
...
    "/database/config": {
      "get": {
        "tags": ["Database"],
        "operationId": "getDbConfig",
        "summary": "Get database configuration",
        "description": "Returns the current database configuration from DatabaseConfig.json (password masked).",
        "responses": {
          "200": { "description": "Database configuration.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      },
      "put": {
        "tags": ["Database"],
        "operationId": "updateDbConfig",
        "summary": "Update database configuration",
        "description": "Persists the provided database configuration to DatabaseConfig.json on the API server's filesystem. Takes effect on next connection pool creation.",
        "requestBody": {
          "required": true,
          "description": "Database configuration to save.",
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DatabaseConfig" } } }
        },
        "responses": {
          "200": { "description": "Config saved to DatabaseConfig.json.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "500": { "description": "File write failed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/system/config": {
      "get": {
        "tags": ["Configuration"],
        "operationId": "listConfig",
        "summary": "List all config entries",
        "description": "Returns all key-value configuration entries from the system_config table, ordered by key. Requires JWT authentication.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "All configuration entries.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfigListResult" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Configuration"],
        "operationId": "upsertConfig",
        "summary": "Create or update config entry",
        "description": "Inserts a new config entry or updates an existing one (upsert on key). Value is stored as JSONB. Requires super_admin role.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "description": "Config entry to create or update.",
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfigEntry" } } }
        },
        "responses": {
          "200": { "description": "Config saved.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "description": "Missing key or value.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Requires super_admin role.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/system/config/{key}": {
      "get": {
        "tags": ["Configuration"],
        "operationId": "getConfig",
        "summary": "Get config by key",
        "description": "Returns a single configuration entry by its key. Requires JWT authentication.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "key", "in": "path", "required": true, "description": "Config key to retrieve", "schema": { "type": "string" }, "example": "auth_provider" }
        ],
        "responses": {
          "200": { "description": "Config entry.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Key not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Configuration"],
        "operationId": "deleteConfig",
        "summary": "Delete config by key",
        "description": "Permanently deletes a configuration entry. Requires super_admin role.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "key", "in": "path", "required": true, "description": "Config key to delete", "schema": { "type": "string" }, "example": "auth_provider" }
        ],
        "responses": {
          "200": { "description": "Config deleted.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Requires super_admin role.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Key not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    }
  }
}